clustering moved to ui utilities as it is a composable
