+++ app/src/main/java/com/monkeyteam/chimpagne/ui/utilities/ChimpagneClustering.kt
@@ -1,4 +1,4 @@
-package com.monkeyteam.chimpagne.model.utils
+package com.monkeyteam.chimpagne.ui.utilities
 
 import android.content.Context
 import androidx.compose.runtime.Composable
